chore: add deriv x trade modal
